pyproject.toml
[tool.poetry]
name = "faridaily-django"
version = "0.1.0"
description = ""
authors = ["frailtynine <frailty.nine@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "5.1.4"
telethon = "^1.38.1"
django-ninja-extra = "^0.22.0"
python-environ = "^0.4.54"
django-cors-headers = "^4.6.0"
django-ninja-jwt = "^5.3.5"
pytest-django = "^4.9.0"
celery = {extras = ["redis"], version = "^5.4.0"}
xlsxwriter = "^3.2.2"
gunicorn = "^23.0.0"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
